Remove packagemanager_aidl from export_shared_lib_headers
Unless a module is exporting libbinder as export_shared_lib_headers,
there is no reason for it to export packagemanager_aidl. Most of them
are consumers of packagemanager_aidl, not providers.

If Removing packagemanager_aidl from export_shared_lib_headers breaks
the build,  then it just means there is some module which is
relying on one of these modules with export_shared_lib_headers as
provider for packagemanager_aidl. We should update them accordingly.

Bug: 183654927
Test: builds + presubmit
Change-Id: Ifc38609b885f4eab55f83d963f6d4d4e295a6791